issue: 1310745 Rename files to lowcase
