+++ src/client.h
@@ -30,7 +30,7 @@
 #define CLIENT_H_
 
 #include "common.h"
-#include "PacketTimes.h"
+#include "packet.h"
 
 //==============================================================================
 //==============================================================================
