+++ src/defs.cpp
@@ -26,37 +26,60 @@
  * OF SUCH DAMAGE.
  *
  */
-#include <stdexcept>
-#include "Defs.h"
-#include "PacketTimes.h"
+#include "defs.h"
+#include "message.h"
+#include "packet.h"
 
-#include "common.h"
+/* Global variables */
+bool g_b_exit = false;
+bool g_b_errorOccured = false;
+uint64_t g_receiveCount = 0; // TODO: should be one per server
+uint64_t g_skipCount = 0; // TODO: should be one per server
 
-PacketTimes::PacketTimes(uint64_t _maxSequenceNo, uint64_t _replyEvery, uint64_t _numServers)
-	: m_maxSequenceNo(_maxSequenceNo)
-	, m_replyEvery(_replyEvery)
-	, m_blockSize(1 + _numServers) // 1 sent + N replies
-	, m_pTimes (new TicksTime[ (_maxSequenceNo / _replyEvery + 1) * m_blockSize])//_maxSequenceNo/_replyEvery+1 is _numBlocks rounded up
-	, m_pInternalUse(&m_pTimes[1])
-	, m_pErrors(new ArrivalErrors[_numServers])
-{
-/*
-	log_msg("m_maxSequenceNo=%lu, m_replyEvery=%lu, m_blockSize=%lu, m_pTimes=%p[%lu], m_pInternalUse=%p, m_pErrors=%p[%lu]"
+unsigned long long g_cycle_wait_loop_counter = 0;
+TicksTime g_cycleStartTime;
 
-			, m_maxSequenceNo, m_replyEvery, m_blockSize, m_pTimes
-			, (_maxSequenceNo / _replyEvery + 1) * m_blockSize
-			, m_pInternalUse
-			, m_pErrors, _numServers
-			);
-//*/
+debug_level_t g_debug_level = LOG_LVL_INFO;
+
+#ifdef  USING_VMA_EXTRA_API
+struct vma_buff_t* g_vma_buff = NULL;
+struct vma_completion_t* g_vma_comps;
+
+ZeroCopyData::ZeroCopyData():
+		m_pkt_buf(NULL),
+		m_pkts(NULL),
+		m_pkt_index(0),
+		m_pkt_offset(0) {
+
+};
+
+void ZeroCopyData::allocate() {
+	m_pkt_buf = (unsigned char *)MALLOC(Message::getMaxSize());
 }
 
-PacketTimes::~PacketTimes() {
-	delete[] m_pTimes;
-	delete[] m_pErrors;
+ZeroCopyData::~ZeroCopyData() {
+	if (m_pkt_buf)
+		FREE(m_pkt_buf);
 }
 
-bool PacketTimes::verifyError(uint64_t _seqNo) {
-	exit_with_err("_seqN > m_maxSequenceNo", SOCKPERF_ERR_FATAL);
-	return false;
-}
+zeroCopyMap g_zeroCopyData;
+#endif
+
+
+uint32_t MPS_MAX = MPS_MAX_UL;     // will be overwrite at runtime in case of ping-pong test
+PacketTimes* g_pPacketTimes = NULL;
+
+TicksTime g_lastTicks;
+
+const int MAX_FDS_NUM = os_get_max_active_fds_num();
+fds_data** g_fds_array = NULL;
+int MAX_PAYLOAD_SIZE = 65507;
+int IGMP_MAX_MEMBERSHIPS = IP_MAX_MEMBERSHIPS;
+
+#ifdef  USING_VMA_EXTRA_API
+struct vma_api_t *g_vma_api;
+#endif
+
+const App* g_pApp = NULL;
+
+
