+++ src/iohandlers.cpp
@@ -26,7 +26,7 @@
  * OF SUCH DAMAGE.
  *
  */
-#include "IoHandlers.h"
+#include "iohandlers.h"
 
 //------------------------------------------------------------------------------
 //------------------------------------------------------------------------------
