+++ src/message.h
