+++ src/playback.h
@@ -30,7 +30,7 @@
 #define __Playback_h__
 
 #include <vector>
-#include "Ticks.h"
+#include "ticks.h"
 
 
 struct PlaybackItem {
